# lib/api-zod/src/generated/types/analytics.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API do TechEra BIOGLOW Tracker
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyticsProblemHistoryItem } from './analyticsProblemHistoryItem';
import type { AnalyticsScoreTrendItem } from './analyticsScoreTrendItem';
import type { MissionMetric } from './missionMetric';

export interface Analytics {
  scoreTrend: AnalyticsScoreTrendItem[];
  missionMetrics: MissionMetric[];
  problemHistory: AnalyticsProblemHistoryItem[];
}