# lib/api-zod/src/generated/types/criterionResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API do TechEra BIOGLOW Tracker
 * OpenAPI spec version: 0.1.0
 */

export interface CriterionResult {
  label: string;
  achieved: boolean;
  /** @minimum 0 */
  quantity: number;
  /** @minimum 0 */
  points: number;
}