# lib/api-zod/src/generated/types/inspection.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API do TechEra BIOGLOW Tracker
 * OpenAPI spec version: 0.1.0
 */
import type { InspectionStatus } from './inspectionStatus';

export interface Inspection {
  status: InspectionStatus;
  /** @minimum 0 */
  points: number;
  notes: string;
}