# lib/api-zod/src/generated/types/memberUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API do TechEra BIOGLOW Tracker
 * OpenAPI spec version: 0.1.0
 */

export interface MemberUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  nickname?: string | null;
  active?: boolean;
}