# lib/api-zod/src/generated/types/missionResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API do TechEra BIOGLOW Tracker
 * OpenAPI spec version: 0.1.0
 */
import type { CriterionResult } from './criterionResult';
import type { MissionResultConfidence } from './missionResultConfidence';
import type { MissionResultStatus } from './missionResultStatus';

export interface MissionResult {
  missionId: number;
  status: MissionResultStatus;
  attempted: boolean;
  /** @minimum 0 */
  points: number;
  criteria: CriterionResult[];
  /** @nullable */
  failureType: string | null;
  technicalNotes: string;
  confidence: MissionResultConfidence;
}