# lib/api-zod/src/generated/types/roundInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API do TechEra BIOGLOW Tracker
 * OpenAPI spec version: 0.1.0
 */
import type { Inspection } from './inspection';
import type { MissionResult } from './missionResult';
import type { RoundInputSeasonName } from './roundInputSeasonName';
import type { RoundInputStatus } from './roundInputStatus';
import type { RoundInputType } from './roundInputType';
import type { Tokens } from './tokens';

export interface RoundInput {
  dateTime: Date;
  type: RoundInputType;
  seasonName?: RoundInputSeasonName;
  event?: string;
  roundNumber?: string;
  memberIds?: number[];
  /** @minimum 1 */
  plannedDurationSeconds: number;
  /**
     * @minimum 0
     * @nullable
     */
  actualDurationSeconds?: number | null;
  robotVersion?: string;
  fieldSetup?: string;
  fieldConditions?: string;
  generalNotes?: string;
  missionResults?: MissionResult[];
  tokens?: Tokens;
  inspection?: Inspection;
  /**
     * @minimum 0
     * @nullable
     */
  officialScore?: number | null;
  officialScoreNotes?: string;
  status?: RoundInputStatus;
}