# lib/api-zod/src/generated/types/team.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API do TechEra BIOGLOW Tracker
 * OpenAPI spec version: 0.1.0
 */
import type { TeamDivision } from './teamDivision';
import type { TeamSeasonName } from './teamSeasonName';

export interface Team {
  id: number;
  name: string;
  number: string;
  city: string;
  country: string;
  seasonName: TeamSeasonName;
  division: TeamDivision;
  robotName: string;
  rulebookVersion: string;
  rulesUpdatedAt: Date;
}