# lib/api-zod/src/generated/types/teamInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API do TechEra BIOGLOW Tracker
 * OpenAPI spec version: 0.1.0
 */

export interface TeamInput {
  /** @minLength 1 */
  name: string;
  number: string;
  city: string;
  country: string;
  robotName: string;
}